Add support for git workspaces in bashrc
Added support for git workspace using cmd. Navigating to git repos now
show the current branch and commit and indicate if there are changes to
be commited. This is implemeted by modifying the bashrc files. Old
my_bashrc also removed.

copy_rcs.sh was also modified to copy the new bash with git support as
well as two auxilary scripts.